add back in old, faster get_runs method..
